// src/interfaces/enum/http.ts
export enum HTTP {
    SUCCESS = 200,
    CREATED = 201,
    ACCEPTED = 202,
    REDIRECT = 301,
    BAD_REQUEST = 400,
    ACCESS_DENIED = 401,
    FORBIDDEN = 403,
    NOT_FOUND = 404,
    NOT_ALLOWED = 405,
    SERVER_ERROR = 500
}

export enum TYPES {
    GET = 'GET',
    POST = 'POST',
    DELETE = 'GELETE'
}